+++ src/components/TagsTable.tsx
@@ -1,41 +1,15 @@
 import { styled } from '@mui/material/styles';
 import { tableCellClasses } from '@mui/material/TableCell';
-import { Box, Stack, Table, TableBody, TableContainer, TableHead, TableRow, Paper, TableCell, Pagination, Typography, TextField, Button } from '@mui/material';
-import { outlinedInputClasses, inputLabelClasses } from '@mui/material';
+import { Box, Stack, Table, TableBody, TableContainer, TableHead, TableRow, Paper, TableCell } from '@mui/material';
 import { indigo, grey } from '@mui/material/colors';
-import { useEffect, useState, ChangeEvent  } from 'react';
 import SortElement from './SortElement';
+import SetResultsElement from './SetResultsElement';
+import ResultsPagination from './ResultsPagination';
 import { Tag, TagArray } from '../types';
 
 function createData(name: string, count: number) {
     return { name, count };
 }
-
-const StyledButton = styled(Button)({
-    backgroundColor: indigo[300],
-    "&:hover": { backgroundColor: indigo[500] }
-})
-
-const StyledTextField = styled(TextField)({
-    [`& .${outlinedInputClasses.root} .${outlinedInputClasses.notchedOutline}`]: {
-        borderColor: indigo[300]
-    },
-    [`&:hover .${outlinedInputClasses.root} .${outlinedInputClasses.notchedOutline}`]: {
-        borderColor: indigo[500]
-    },
-    [`& .${outlinedInputClasses.root}.${outlinedInputClasses.focused} .${outlinedInputClasses.notchedOutline}`]: {
-        borderColor: indigo[500]
-    },
-    [`& .${inputLabelClasses.outlined}`]: {
-        color: indigo[300]
-    },
-    [`&:hover .${inputLabelClasses.outlined}`]: {
-        color: indigo[500]
-    },
-    [`& .${inputLabelClasses.outlined}.${inputLabelClasses.focused}`]: {
-        color: indigo[500]
-    }
-})
 
 const StyledTableCell = styled(TableCell)({
     [`&.${tableCellClasses.head}`]: {
@@ -51,18 +25,6 @@ const StyledTableRow = styled(TableRow)({
     }
 })
 
-const StyledPagination = styled(Pagination)({
-    marginTop: "1rem",
-    justifyContent: "center",
-    display: 'flex',
-    "& .MuiPaginationItem-root:hover": {
-        backgroundColor: indigo[400],
-    },
-    "& .Mui-selected": {
-        backgroundColor: indigo[200],
-    }
-})
-
 type Props = {
     tags: TagArray,
     totalPages: number | undefined,
@@ -73,48 +35,12 @@ type Props = {
 }
 
 const TagsTable = ({ tags, setOrder, setSort, setPagesize, totalPages, setPage }: Props) => {
-    const [resultsNumber, setResultsNumber] = useState(10)
-    const [inputError, setInputError] = useState('')
 
     const rows = tags.map((el: Tag) => createData(el.name, el.count))
 
-    const handleSetPagesize = () => {
-        setInputError('')
-        if (resultsNumber > 0) setPagesize(resultsNumber)
-        else setInputError('Results number must be greater than 0')
-    }
-
-    useEffect(() => {
-        console.log('results number', resultsNumber)
-    }, [resultsNumber])
-
-    const handlePageChange = (event: ChangeEvent<unknown>, value: number) => {
-        setPage(value)
-    }
-
     return (
         <>
-            <Stack sx={{ marginBottom: "1rem" }}>
-                <Stack direction='row' alignItems='center' gap={2} sx={{ marginBottom: ".25rem" }}>
-                    <StyledTextField type="number" id="outlined-number" label="Results per page" size="small" InputLabelProps={{
-                        shrink: true,
-                    }}
-                        InputProps={{
-                            inputProps: {
-                                min: 1
-                            }
-                        }}
-                        sx={{ width: 200, fieldset: { borderColor: indigo[400] } }} defaultValue={10}
-                        onChange={(event) => {
-                            console.log(event.target.value)
-                            console.log('number?', typeof (event.target.value))
-                            setResultsNumber(Number(event.target.value))
-                        }}
-                    />
-                    <StyledButton variant="contained" onClick={handleSetPagesize}>Set pages</StyledButton>
-                </Stack>
-                {<Typography sx={{ color: 'error.main' }}>{inputError}</Typography>}
-            </Stack>
+            <SetResultsElement setPagesize={setPagesize} />
             <TableContainer component={Paper}>
                 <Table>
                     <TableHead >
@@ -143,7 +69,7 @@ const TagsTable = ({ tags, setOrder, setSort, setPagesize, totalPages, setPage }
                     </TableBody>
                 </Table>
             </TableContainer>
-            <StyledPagination count={totalPages} sx={{ color: indigo[500] }} onChange={handlePageChange} />
+            <ResultsPagination totalPages={totalPages} setPage={setPage} />
         </>
     )
 }
